backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.1
openai>=1.3.0
